refactor: collapse listeningHistory three-hop chain
Before:
  utils/listeningHistory.ts
    → re-exports from db/listeningHistory.ts (relative)
      → instantiates createListeningHistoryRepo from @rainbot/db
        via the relative path ../packages/db/src/listeningHistory
  apps/raincloud/db/listeningHistory.ts (re-exports from db/listeningHistory.ts, unused)

After:
  utils/listeningHistory.ts
    → imports createListeningHistoryRepo from @rainbot/db,
      instantiates the singleton, exports the bound methods.

The audit framed this as "three-hop re-export chain" but the middle
hop (db/listeningHistory.ts) actually owned the singleton wiring
(query, logger, detectSourceType injection). Moved the wiring into
utils/listeningHistory.ts since it's already the consumer-facing
module, and dropped the intermediate file.

apps/raincloud/db/listeningHistory.ts had zero consumers — deleted.

Added @rainbot/db to apps/raincloud and apps/pranjeet deps so the
@rainbot/db import resolves in workspaces that compile utils/.

type-check passes across all 18 tasks.
