+++ apps/pranjeet/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@discordjs/opus": "^0.10.0",
     "@discordjs/voice": "^0.19.0",
+    "@rainbot/db": "*",
     "@rainbot/protocol": "*",
     "@rainbot/rpc": "*",
     "@rainbot/shared": "*",
