--- apps/raincloud/db/listeningHistory.ts
@@ -1,1 +1,0 @@
-export * from '../../db/listeningHistory';
