--- db/listeningHistory.ts
@@ -1,29 +1,0 @@
-import { createListeningHistoryRepo } from '../packages/db/src/listeningHistory';
-import { createLogger } from '../utils/logger';
-import { query } from '../utils/database';
-import { detectSourceType } from '../utils/sourceType';
-
-const log = createLogger('HISTORY');
-
-const listeningHistoryRepo = createListeningHistoryRepo({
-  query,
-  logger: log,
-  detectSourceType,
-});
-
-export const {
-  saveHistory,
-  getHistory,
-  clearHistory,
-  trackPlayed,
-  getListeningHistory,
-  getRecentHistory,
-  clearListeningHistory,
-} = listeningHistoryRepo;
-
-export type {
-  HistoryEntry,
-  ListeningHistoryRow,
-  Track,
-  SourceType,
-} from '../packages/db/src/listeningHistory';
